LPS-124898 There is strong coupling between OfflineOpenIdConnectSessionManager and OpenIdConnectTokenRefreshMessageListener so just combine these and avoid unnecessary public methods
