+++ modules/apps/portal-security-sso/portal-security-sso-openid-connect-impl/src/main/java/com/liferay/portal/security/sso/openid/connect/internal/session/manager/OfflineOpenIdConnectSessionManager.java
@@ -19,10 +19,22 @@ import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.messaging.BaseMessageListener;
+import com.liferay.portal.kernel.messaging.Destination;
+import com.liferay.portal.kernel.messaging.DestinationConfiguration;
+import com.liferay.portal.kernel.messaging.DestinationFactory;
+import com.liferay.portal.kernel.messaging.Message;
+import com.liferay.portal.kernel.messaging.MessageListener;
 import com.liferay.portal.kernel.scheduler.SchedulerException;
+import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.Time;
+import com.liferay.portal.security.sso.openid.connect.OpenIdConnectProviderRegistry;
 import com.liferay.portal.security.sso.openid.connect.constants.OpenIdConnectWebKeys;
+import com.liferay.portal.security.sso.openid.connect.internal.constants.OpenIdConnectDestinationNames;
 import com.liferay.portal.security.sso.openid.connect.internal.scheduler.OpenIdConnectTokenRefreshScheduler;
+import com.liferay.portal.security.sso.openid.connect.internal.util.OpenIdConnectTokenRequestUtil;
 import com.liferay.portal.security.sso.openid.connect.persistence.model.OpenIdConnectSession;
 import com.liferay.portal.security.sso.openid.connect.persistence.service.OpenIdConnectSessionLocalService;
 import com.liferay.portal.security.sso.openid.connect.session.manager.OpenIdConnectSessionManager;
@@ -31,12 +43,18 @@ import com.nimbusds.oauth2.sdk.ParseException;
 import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.oauth2.sdk.token.RefreshToken;
 import com.nimbusds.oauth2.sdk.util.JSONObjectUtils;
+import com.nimbusds.openid.connect.sdk.op.OIDCProviderMetadata;
+import com.nimbusds.openid.connect.sdk.rp.OIDCClientMetadata;
 import com.nimbusds.openid.connect.sdk.token.OIDCTokens;
 
 import java.util.Date;
+import java.util.Dictionary;
 
 import javax.servlet.http.HttpSession;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
@@ -66,38 +84,6 @@ public class OfflineOpenIdConnectSessionManager
 		catch (SchedulerException schedulerException) {
 			if (_log.isWarnEnabled()) {
 				_log.warn(schedulerException, schedulerException);
-			}
-		}
-	}
-
-	public void extendOpenIdConnectSession(
-			long openIdConnectSessionId, OIDCTokens oidcTokens)
-		throws PortalException {
-
-		OpenIdConnectSession openIdConnectSession =
-			_openIdConnectSessionLocalService.getOpenIdConnectSession(
-				openIdConnectSessionId);
-
-		AccessToken oldAccessToken = _getAccessToken(openIdConnectSession);
-
-		AccessToken accessToken = oidcTokens.getAccessToken();
-
-		_updateOpenIdConnectSession(
-			accessToken, oidcTokens.getRefreshToken(), openIdConnectSession);
-
-		if ((openIdConnectSession.getRefreshToken() != null) &&
-			(oldAccessToken.getLifetime() != accessToken.getLifetime())) {
-
-			try {
-				_openIdConnectTokenRefreshScheduler.reschedule(
-					accessToken.getLifetime(),
-					openIdConnectSession.getModifiedDate(),
-					openIdConnectSession.getOpenIdConnectSessionId());
-			}
-			catch (SchedulerException schedulerException) {
-				if (_log.isWarnEnabled()) {
-					_log.warn(schedulerException, schedulerException);
-				}
 			}
 		}
 	}
@@ -181,14 +167,94 @@ public class OfflineOpenIdConnectSessionManager
 		return openIdConnectSession.getOpenIdConnectSessionId();
 	}
 
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_bundleContext = bundleContext;
+
+		DestinationConfiguration destinationConfiguration =
+			DestinationConfiguration.createSerialDestinationConfiguration(
+				OpenIdConnectDestinationNames.OPENID_CONNECT_TOKEN_REFRESH);
+
+		Destination destination = _destinationFactory.createDestination(
+			destinationConfiguration);
+
+		Dictionary<String, Object> dictionary =
+			HashMapDictionaryBuilder.<String, Object>put(
+				"destination.name", destination.getName()
+			).build();
+
+		_serviceRegistration1 = bundleContext.registerService(
+			Destination.class, destination, dictionary);
+
+		_serviceRegistration2 = bundleContext.registerService(
+			MessageListener.class, new OpenIdConnectMessageListener(),
+			dictionary);
+	}
+
 	@Deactivate
 	protected void deactivate() throws Exception {
+		if (_serviceRegistration1 != null) {
+			Destination destination = _bundleContext.getService(
+				_serviceRegistration1.getReference());
+
+			_serviceRegistration1.unregister();
+
+			destination.destroy();
+		}
+
+		if (_serviceRegistration2 != null) {
+			_serviceRegistration2.unregister();
+		}
+
+		_bundleContext = null;
+
 		for (OpenIdConnectSession openIdConnectSession :
 				_openIdConnectSessionLocalService.getOpenIdConnectSessions(
 					QueryUtil.ALL_POS, QueryUtil.ALL_POS)) {
 
 			_openIdConnectSessionLocalService.deleteOpenIdConnectSession(
 				openIdConnectSession);
+		}
+	}
+
+	private void _extendOpenIdConnectSession(
+			OpenIdConnectSession openIdConnectSession)
+		throws Exception {
+
+		if (openIdConnectSession.getRefreshToken() == null) {
+			return;
+		}
+
+		RefreshToken refreshToken = new RefreshToken(
+			openIdConnectSession.getRefreshToken());
+
+		OIDCTokens oidcTokens = OpenIdConnectTokenRequestUtil.request(
+			_openIdConnectProviderRegistry.findOpenIdConnectProvider(
+				CompanyThreadLocal.getCompanyId(),
+				openIdConnectSession.getProviderName()),
+			refreshToken);
+
+		AccessToken oldAccessToken = _getAccessToken(openIdConnectSession);
+
+		AccessToken accessToken = oidcTokens.getAccessToken();
+
+		_updateOpenIdConnectSession(
+			accessToken, oidcTokens.getRefreshToken(), openIdConnectSession);
+
+		if ((openIdConnectSession.getRefreshToken() != null) &&
+			(oldAccessToken.getLifetime() != accessToken.getLifetime())) {
+
+			try {
+				_openIdConnectTokenRefreshScheduler.reschedule(
+					accessToken.getLifetime(),
+					openIdConnectSession.getModifiedDate(),
+					openIdConnectSession.getOpenIdConnectSessionId());
+			}
+			catch (SchedulerException schedulerException) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(schedulerException, schedulerException);
+				}
+			}
 		}
 	}
 
@@ -240,8 +306,18 @@ public class OfflineOpenIdConnectSessionManager
 	private static final Log _log = LogFactoryUtil.getLog(
 		OfflineOpenIdConnectSessionManager.class);
 
+	private volatile BundleContext _bundleContext;
+
 	@Reference
 	private CounterLocalService _counterLocalService;
+
+	@Reference
+	private DestinationFactory _destinationFactory;
+
+	@Reference
+	private OpenIdConnectProviderRegistry
+		<OIDCClientMetadata, OIDCProviderMetadata>
+			_openIdConnectProviderRegistry;
 
 	@Reference
 	private OpenIdConnectSessionLocalService _openIdConnectSessionLocalService;
@@ -249,5 +325,31 @@ public class OfflineOpenIdConnectSessionManager
 	@Reference
 	private OpenIdConnectTokenRefreshScheduler
 		_openIdConnectTokenRefreshScheduler;
+
+	private ServiceRegistration<Destination> _serviceRegistration1;
+	private ServiceRegistration<MessageListener> _serviceRegistration2;
+
+	private class OpenIdConnectMessageListener extends BaseMessageListener {
+
+		protected void doReceive(Message message) throws Exception {
+			long openIdConnectSessionId = GetterUtil.getLong(
+				message.getPayload());
+
+			try {
+				_extendOpenIdConnectSession(
+					_openIdConnectSessionLocalService.getOpenIdConnectSession(
+						openIdConnectSessionId));
+			}
+			catch (PortalException portalException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(
+						"Unable to find OpenIdConnectSession with Id " +
+							openIdConnectSessionId,
+						portalException);
+				}
+			}
+		}
+
+	}
 
 }
